+++ integration_tests/suite/test_phone_view.py
@@ -159,12 +159,6 @@ class TestPhoneView(BaseDirdIntegrationTest):
         )
         assert_that(result.status_code, equal_to((200)))
 
-    def test_yealink_lookup_route(self):
-        result = self.get_lookup_yealink_result(
-            self.profile, VALID_UUID, term='Alice', token=VALID_TOKEN_MAIN_TENANT
-        )
-        assert_that(result.status_code, equal_to((200)))
-
     def test_gigaset_lookup_route(self):
         result = self.get_lookup_gigaset_result(
             self.profile, VALID_UUID, term='Alice', token=VALID_TOKEN_MAIN_TENANT